Renaming components and moving styling sheets to their own folder
