+++ src/styles/poiCard.module.css
